backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[[tool.poetry.source]]
name = "ali"
default = true
url = "https://mirrors.aliyun.com/pypi/simple/"


[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "3.13.1"
django = "4.0.1"
flake8 = "4.0.1"
gunicorn = "20.0.4"
uvicorn = "^0.17.0"
django-cors-headers = "^3.11.0"
drf-yasg = "^1.20.0"
channels = "^3.0.4"
channels-redis = "^3.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
